add claim loader and jwt optional to the project

+++ resources/item.py
@@ -1,7 +1,6 @@
-from flask_jwt_extended import jwt_required
+from flask_jwt_extended import jwt_required, get_jwt_claims, jwt_optional, get_jwt_identity
 from flask_restful import Resource, reqparse
 from models.item import ItemModel
-
 
 class Item(Resource):
     parser = reqparse.RequestParser()
@@ -35,7 +34,12 @@ class Item(Resource):
 
        return item.json(), 201
 
+    @jwt_required
     def delete(self, name):
+        claims = get_jwt_claims()
+        if not claims['is_admin']:
+            return {'msg': 'admin privilage required!'}, 401
+
         item = ItemModel.find_by_name(name)
         if item:
             item.delete_from_db()
@@ -57,5 +61,12 @@ class Item(Resource):
 
 
 class ItemList(Resource):
+    @jwt_optional
     def get(self):
-        return{'items': list(map(lambda x:x.json(), ItemModel.find_all))}
+        user_id = get_jwt_identity()
+        items = [item.json() for item in ItemModel.find_all()]
+        if user_id:
+            return {'items': items}
+        return {'items': [item['name'] for item in items],
+                'msg': 'More information available if you login'}, 200
+
